+++ aloe/testclass.py
@@ -309,7 +309,7 @@ def run_example(self):
         in the feature file.
         """
 
-        assert len(steps) > 0
+        assert steps
 
         step_definitions = [
             cls.prepare_step(step)